Lights and lightmap added
Lights (light_t structure) and whole-world lightmap was added with this commit. The lightmap is calculated from the 2D grid (although the lights are 3D points because at some point in the future they will be used to calculate a generic "light direction" at every cluster which will be used when rendering entities) and is always bound at texture unit 1 (GL_TEXTURE1). The entities are being rendered using the usual vertical directional light like before, but the diffuse and ambient colors are set to the lightmap value at the cell where the entity's origin is located (the ambient color is set to 1/3 of the diffuse color - maybe later a light can define an ambient color and the clusters will average the ambient colors which will be used for that).