Cluster mesh optimization added
Now the cluster mesh is optimized to avoid unneccessary quads inside the mesh (still there is at least one quad per cluster, but that is ok). This optimization increased the speed at One's GMA950 rendering to more than 30fps even in the case where almost the whole map is visible.

Still, the entities take a considerable amount of processing power. But this will -hopefully- be solved when LOD is added for them.