Improved GUI code, added checkbox, finished field
The GUI code is improved to support clipping, added checkboxes and finished the field. The checkbox can act as a normal checkbox or as a radiobutton based on the value of its 'group' field. If the group is zero, the control acts as a checkbox, otherwise it acts as a radiobutton.

The field is finished to be fully usable and can use arrow keys and home/end buttons.

The font rendering has been improved to use screen ratio in empty characters.