Resizeable windows, gui fix and ray_march function
This commit adds resizeable windows in the GUI, fixes a clipping issue and adds a generic ray marching function for the world. The rendering still uses a special ray marching function to avoid cache misses, but the new one can be used when a few rays needs to be shot in the world.