Removed LIL source code in order to use liblil.a
This commit removes the LIL source code and adds a dependency on liblil.a. The configure.py script looks for lil in ../lil, ../../lil and ./lil (in that order) and if none is found, it expects from the user to specify a --lil-path option with the path to liblil.a and lil.h.