Improved motion, ai, scripting, etc
This commit improves motion code for better "on-ground" checking. Also adds a few new functions for scripting, like the shortcut script_set_int/string/float set of functions which set a global variable using an int/string/float value.

It also improves a lot the AI scripting support by fixing movement and adding lil callbacks via entity attributes. The attributes hold lil code and are called depending on an event mask (in entities) which is updated automatically when the attribute value is changed. The event attributes so far are:

 * motion-touch, which is called when the entity touches another entity.
 * motion-touched, which is the opposite of motion-touch and called when the entity is touched by another entity. Note that this event is actually sent by the motion controller of the other entity so the touched entity doesn't have to have motion initialized (ie. it can be received by motionless/static entities). This is sent after motion-touch.
 * motion-hit-ground, which is called when the entity hits the ground.
 * move-target-reached, which is called when the target set via ai_move_to/move-entity-to is reached.

A lesser change is that the default world now uses new textures and the ceiling and upper texture is fully a skybox view. Also the ceiling is much taller giving the impression of a walled area and allowing the throwing of cameramen much higher.

Also updated .gitignore to ignore some extra files.

Note: this commit requires the data files from alithia-data-2010-08-23.tar.xz