Skybox rendering now uses stencil to avoid gaps
This commit makes skybox rendering to use the stencil buffer to draw only the areas covered by skybox polygons in order to avoid gaps introduced by inaccuracies in visibility rays and t-junctions. Also thanks to this commit the skybox overdraw (which initially covered the whole screen only to be replaced by the world geometry, minus the skybox polygon parts) is avoided.