Modified visibility marking code
This commit modifies the visibility marking code in proc_map to mark the clusters of the cells *around* the cell in question instead of the cluster of the cell itself and the cluster next to it. This produces both more accurate results and less visible clusters. However there are more cache misses which makes FPSs in large areas (like the default open room) to be much slower.