Added entitylevel textures, shadow flag and bugfix
This commit adds entity-level textures which can be used to render a model using some world texture (with automatic box projected UVs) as part of the world geometry.

Currently the assignment is hardcoded and the world format doesn't contain these entity-level textures.

Also fixes a small bug with world saving and adds a "draw-shadow" flag to entities (set using entity attributes) which now by default is not set.