Added cell offset reseting functions and some minor other stuff
This commit adds two new script functions, reset-cell-floor-offsets and reset-cell-ceiling-offsets which reset (set to 0) the offsets for floor and ceiling of a cell.

Also makes sure the camera rotation is proper and adds code which releases the mouse when the editor is in GUI mode and fullscreen is set to 0.